refactor: rename constraints to rule_core
