+++ crates/config/src/rule_core.rs
